+++ src/collective/cover/tiles/embed.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
-
 from collective.cover import _
 from collective.cover.tiles.base import IPersistentCoverTile
 from collective.cover.tiles.base import PersistentCoverTile
+from plone.autoform.directives import write_permission
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope import schema
 from zope.interface import implements
@@ -10,6 +10,7 @@ from zope.interface import implements
 
 class IEmbedTile(IPersistentCoverTile):
 
+    write_permission(embed='collective.cover.EmbedCode')
     embed = schema.Text(
         title=_(u'Embedding code'),
         required=False,